+++ clang/docs/ReleaseNotes.rst
@@ -310,6 +310,8 @@ AST Matchers
 
 clang-format
 ------------
+- Add ``NumericLiteralCase`` option for enforcing character case in numeric
+  literals.
 
 libclang
 --------